requirements.txt
mcp>=1.0.0
rank-bm25>=0.2.2
httpx>=0.27.0
pdfplumber>=0.11.0
tqdm>=4.66.0
python-dotenv>=1.0.0
google-genai>=1.0.0
rich>=13.7.0
aiofiles>=23.2.0
tiktoken>=0.7.0
pydantic>=2.7.0
qdrant-client>=1.9.0
fastapi>=0.111.0
uvicorn>=0.30.0
supabase>=2.0.0
